requirements.txt
streamlit==1.32.0
gspread==6.0.2
oauth2client==4.1.3
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1 
